Amalgamating T model function modules
